requirements/local.txt
EasyProcess==0.2.3
numpy==1.14.2
olefile==0.44
Pillow==4.3.0
pyPdf==1.13
PyPDF2==1.26.0
pyscreenshot==0.4.2